project binance/antrade-main/pyproject.toml
[tool.poetry]
name = "antrade"
version = "0.1.0"
description = "Platform for algorithmic trading"
authors = ["jardinesdelareina <jardinesdelareina@mail.ru>"]
license = "BSD 3-Clause License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.1"
python-binance = "^1.0.17"
environs = "^9.5.0"
pyclean = "^2.7.2"
aiogram = "^2.25.1"
plotly = "^5.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
